deleteProject updated

+++ src/main/java/com/lespania/implementation/ProjectServiceImpl.java
@@ -84,11 +84,17 @@ public class ProjectServiceImpl implements ProjectService {
     }
 
     @Override
-    public void delete(String code) {
+    public void delete(String code) throws TicketingProjectException {
         Project project = projectRepository.findByProjectCode(code);
+
+        if(project == null){
+            throw new TicketingProjectException("Project does not exist");
+        }
+
         project.setIsDeleted(true);
 
         project.setProjectCode(project.getProjectCode() +  "-" + project.getId());
+
         projectRepository.save(project);
 
         taskService.deleteByProject(mapperUtil.convert(project,new ProjectDTO()));
